Extract AbstractDeltaLakePageSink from DeltaLakePageSink
